Use write/readback to verify 4bit mode on Macronix MX25L3233F.
The old macro used QPI commands which are not supported by the
Macronix MX25L3233F device. Now a pattern is written to the last
flash sector to provide well-known data for a 4 bit read operation.
The old contents of the flash sector are not restored, which makes
this test destructive for existing flash contents.